refactor: extract update flow from main.rs into update_cmd module
Moves the update/post-update wiring (plugin refresh, daemon refresh,
subprocess re-exec, health pass) into src/update_cmd.rs following the
*_cmd convention, bringing main.rs to 871 lines. Also promotes the
branch-DB tests' IsolatedEnv into tests/common as the canonical
env-isolation helper.
